Working custom particle shape
creating the entire cube out of moveable vertices was becoming a headache so I took one working pyramid shape and then just repeated it 6 times while rotating it to get the desired shape. The next step is to try and fit it into a PShape group object then create an array of them.