a1 readme update
I've been using a mouse based rotation system that is always active to rotate the test particle(s) and now I changed it so you have to click and drag to activate the system which feels more natural than having it always track the mouse; however it is kind of jumpy currently so towards the end of the project I will tweak it to feel smoother. 

It also just occurred to me that this could probably count as an extra feature.