New dynamic shape system
I should have probably done this from the start but I converted over to a point system where the corners of the cubes are put in and the faces of the cube are then generated. I was kind of forced to swap to this system as drawing one face made me encounter a bug deep within processing's 3D renderer. Rotating shapes that are a group of shapes on the Z axis fails and throws compile errors. It is a reported bug but has yet to be fixed. So now I generate each face of the cube separately instead of drawing one face and rotating it around to create the cube.